+++ Hearthstone Deck Tracker/Hearthstone/CounterSystem/Counters/DiscoversThisGameCounter.cs
@@ -1,0 +1,54 @@
+﻿using HearthDb.Enums;
+using Hearthstone_Deck_Tracker.LogReader.Interfaces;
+using Hearthstone_Deck_Tracker.Utility;
+using Entity = Hearthstone_Deck_Tracker.Hearthstone.Entities.Entity;
+
+namespace Hearthstone_Deck_Tracker.Hearthstone.CounterSystem.Counters;
+
+public class DiscoversThisGameCounter : NumericCounter
+{
+	protected override string? CardIdToShowInUI => HearthDb.CardIds.Collectible.Hunter.AlienEncounters;
+
+	public override string[] RelatedCards => new string[]
+	{
+		HearthDb.CardIds.Collectible.Hunter.AlienEncounters,
+	};
+
+	public DiscoversThisGameCounter(bool controlledByPlayer, GameV2 game) : base(controlledByPlayer, game)
+	{
+	}
+
+	public override bool ShouldShow()
+	{
+		if(!Game.IsTraditionalHearthstoneMatch) return false;
+		if(IsPlayerCounter)
+			return InPlayerDeckOrKnown(RelatedCards);
+		return Counter > 1 && OpponentMayHaveRelevantCards();
+	}
+
+	public override string[] GetCardsToDisplay()
+	{
+		return IsPlayerCounter ?
+			GetCardsInDeckOrKnown(RelatedCards).ToArray() :
+			FilterCardsByClassAndFormat(RelatedCards, Game.Opponent.OriginalClass);
+	}
+
+	public override string ValueToShow() => Counter.ToString();
+
+	public override void HandleChoicePicked(IHsCompletedChoice choice)
+	{
+		if(!Game.IsTraditionalHearthstoneMatch)
+			return;
+
+		if(!Game.Entities.TryGetValue(choice.SourceEntityId, out var source))
+			return;
+
+		if(source.GetTag(GameTag.DISCOVER) == 0)
+			return;
+
+		var controller = source.GetTag(GameTag.CONTROLLER);
+
+		if((controller == Game.Player.Id && IsPlayerCounter) || (controller == Game.Opponent.Id && !IsPlayerCounter))
+			Counter++;
+	}
+}